Added Login Testing Report
This report depicts 9 combinations for valid and non valid Clients attempting to login to our website. These inputs should be exhaustive to ensure the quality of our website.

The report describes success and failure scenarios for these login attempts on the Chrome browser, as well as the time it took to determine that outcome.