Added Promo Code Testing Report
This report depicts 3 combinations for valid and non valid and blank promo codes. These inputs should be exhaustive to ensure the quality of our website.

The report describes success and failure scenarios for these input promo code attempts on the Chrome browser, as well as the time it took to determine that outcome.